:sparkles: run on post
