+++ action.yml
@@ -14,4 +14,6 @@ inputs:
     description: 'todo'
 runs:
   using: 'node12'
-  main: 'dist/index.js'
+  main: 'dist/main/index.js'
+  post: 'dist/post/index.js'
+  post-if: 'success()'
