+++ src/main.ts
@@ -1,18 +1,14 @@
 import * as core from '@actions/core'
-import {exec, ExecOptions} from '@actions/exec'
+import {exec} from '@actions/exec'
 
 async function run(): Promise<void> {
   try {
     const token: string = core.getInput('github-token')
-    const dir: string = core.getInput('working-directory')
-    const flag: string = core.getInput('flag-name')
-
     if (!token) {
       throw new Error("'github-token' input missing")
     }
 
     await installDep()
-    await report(token, flag, dir)
   } catch (error) {
     core.setFailed(error.message)
   }
@@ -20,28 +16,6 @@ async function run(): Promise<void> {
 
 async function installDep(): Promise<void> {
   await exec('cpanm', ['-n', 'Devel::Cover::Report::Coveralls'])
-}
-
-async function report(
-  token: string,
-  flag: string | undefined,
-  dir: string | undefined
-): Promise<void> {
-  const opts: ExecOptions = {
-    env: {
-      GITHUB_TOKEN: token
-    }
-  }
-
-  if (dir) {
-    opts.cwd = dir
-  }
-
-  if (flag && opts.env) {
-    opts.env['COVERALLS_FLAG_NAME'] = flag
-  }
-
-  await exec('cover', ['-report', 'coveralls'], opts)
 }
 
 run()
